move site to render

+++ reset_db.py
@@ -1,9 +1,20 @@
-import os
 from db import engine, Base
-from models import User, PaymentTransaction
 
-print("Dropping all tables...")
-Base.metadata.drop_all(bind=engine)
-print("Creating all tables...")
-Base.metadata.create_all(bind=engine)
-print("Database reset successful.")
+def main():
+	confirm = input("It will delete all from Database! Contineu? (yes/NO): ")
+
+	if confirm.lower() == "yes":
+		print("Dropping all tables...")
+		Base.metadata.drop_all(bind=engine)
+
+		print("Creating all tables...")
+		Base.metadata.create_all(bind=engine)
+
+		print("Database reset successful.")
+
+	else:
+		print("Aborting...")
+
+
+if __name__ == '__main__':
+	main()
